rename/move pki package to pkg/pki
